backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
drf-spectacular==0.28.0
environ==1.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.0
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
